Impersonate address needs to return the managed signer (#503)
