+++ scenario/context/CometContext.ts
@@ -14,13 +14,23 @@ import {
 } from '../constraints';
 import CometActor from './CometActor';
 import CometAsset from './CometAsset';
-import { Comet, CometInterface, ProxyAdmin, ERC20, ERC20__factory, Configurator, SimpleTimelock, CometProxyAdmin, GovernorSimple } from '../../build/types';
+import {
+  Comet,
+  CometInterface,
+  ProxyAdmin,
+  ERC20,
+  ERC20__factory,
+  Configurator,
+  SimpleTimelock,
+  CometProxyAdmin,
+  GovernorSimple,
+  IGovernorBravo,
+} from '../../build/types';
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
 import { sourceTokens } from '../../plugins/scenario/utils/TokenSourcer';
 import { ProtocolConfiguration, deployComet } from '../../src/deploy';
 import { AddressLike, getAddressFromNumber, resolveAddress } from './Address';
 import { Requirements } from '../constraints/Requirements';
-import { fastGovernanceExecute } from '../utils';
 
 type ActorMap = { [name: string]: CometActor };
 type AssetMap = { [name: string]: CometAsset };
@@ -231,10 +241,72 @@ export class CometContext {
     await this.world.hre.network.provider.send('hardhat_setNextBlockBaseFeePerGas', ['0x0']);
   }
 
+  async setNextBlockTimestamp(timestamp: number) {
+    await this.world.hre.ethers.provider.send('evm_setNextBlockTimestamp', [timestamp]);
+  }
+
+  // XXX Hardhat 2.9 has 'hardhat_mine' for mining multiple blocks
+  async mineBlocks(blocks: number) {
+    for (let i = 0; i < blocks; i++) {
+      await this.world.hre.network.provider.send('evm_mine', []);
+    }
+  }
+
   // Instantly executes some actions through the governance proposal process
+  // Note: `governor` must be connected to an `admin` signer
   async fastGovernanceExecute(targets: string[], values: BigNumberish[], signatures: string[], calldatas: string[]) {
-    let governor = await this.getGovernor();
-    await fastGovernanceExecute(this.world, governor, targets, values, signatures, calldatas);
+    const { world } = this;
+    const governor = await this.getGovernor();
+    // XXX See if there is a better way to determine if GovernorSimple or Bravo should be used
+    try {
+      const admin = await governor.admins(0);
+      const adminSigner = await world.impersonateAddress(admin);
+      const governorAsAdmin = governor.connect(adminSigner);
+
+      const tx = await (await governorAsAdmin.propose(targets, values, signatures, calldatas, 'FastExecuteProposal')).wait();
+      const event = tx.events.find(event => event.event === 'ProposalCreated');
+      const [proposalId] = event.args;
+
+      await governorAsAdmin.queue(proposalId);
+      await governorAsAdmin.execute(proposalId);
+    } catch (e) {
+      // XXX find a better way to do this without hardcoding whales
+      const voters = [
+        '0xea6c3db2e7fca00ea9d7211a03e83f568fc13bf7',
+        '0x683a4f9915d6216f73d6df50151725036bd26c02'
+      ];
+      const adminSigner = await world.impersonateAddress(voters[0]);
+      const governorAsAdmin = await world.hre.ethers.getContractAt(
+        'IGovernorBravo',
+        governor.address,
+        adminSigner
+      ) as IGovernorBravo;
+
+      const proposeTxn = await (await governorAsAdmin.propose(targets, values, signatures, calldatas, 'FastExecuteProposal')).wait();
+      const proposeEvent = proposeTxn.events.find(event => event.event === 'ProposalCreated');
+      const [proposalId, , , , , , startBlock, endBlock] = proposeEvent.args;
+
+      const blocksUntilStart = startBlock - await world.hre.ethers.provider.getBlockNumber();
+      const blocksFromStartToEnd = endBlock - startBlock;
+      await this.mineBlocks(blocksUntilStart);
+      for (const voter of voters) {
+        const voterSigner = await world.impersonateAddress(voter);
+        const govAsVoter = await world.hre.ethers.getContractAt(
+          'IGovernorBravo',
+          governor.address,
+          voterSigner
+        ) as IGovernorBravo;
+        await govAsVoter.castVote(proposalId, 1);
+      }
+
+      await this.mineBlocks(blocksFromStartToEnd);
+      const queueTxn = await (await governorAsAdmin.queue(proposalId)).wait();
+      const queueEvent = queueTxn.events.find(event => event.event === 'ProposalQueued');
+      let [proposalId_, eta] = queueEvent.args;
+
+      await this.setNextBlockTimestamp(eta.toNumber());
+      await governorAsAdmin.execute(proposalId);
+    }
   }
 }
